Add pro forma tests
This commit adds a pro forma unit test to say we have working testing.
This will be expanded in the future as the functionality of the tool
expands. For now it's still early and testing doesn't buy us much (since
it doesn't really do anything). This is mostly to make sure we have a
working test running infrastructure setup.