Make create_image call functional
This commit fixes a couple of issues with the create_image function
found from actually trying to use it. The first issue is despite the
python client docs indicating the os_code is an optional parameter, it
turns out it is required. (which matches the web ui experience) The
second is the swift username for the external uri is not your username
but actually the project name.