Expand the README
This commit expands the readme to document how to get the current state
of the project running. It will get you as far as I've currently gotten
which is uploading an image and booting it. However the current state
still doesn't work 100% as noted in the readme. Hopefully someone else
will pick this up and take it the rest of the way.